Rename `extSPromise.ts` to `haven-storage.ts`
